Draft: move functions to draftgeoutils.wires
