+++ src/Mod/Draft/CMakeLists.txt
@@ -36,6 +36,7 @@ SET (Draft_geoutils
     draftgeoutils/sort_edges.py
     draftgeoutils/faces.py
     draftgeoutils/geometry.py
+    draftgeoutils/wires.py
 )
 
 SET(Draft_tests
